Bug 1294649 - Don't show EME plugins in the plugin list that are disabled. r=spohl
I noticed that on Linux is you have 'Play DRM content' unchecked, and then
open about:addons > plugins, and then check 'Play DRM content', that the
Adobe CDM then appears in the about:addons > plugins list.

This is because in GMPWrapper.onPrefEMEGlobalEnabledChanged() we're taking the
branch that calls AddonManagerPrivate addon listeners.

We're doing that because we're only checking the GMPWrapper.appDisabled on
that branch, and so if EME is enabled, we'll go down that path for the Adobe
CDM on Linux, even though the CDM itself isn't visible/supported on Linux.

So we should instead only go down that path if EME is enabled, *and* if the
GMP is not hidden. We can uninstall the GMP if it's hidden or EME is
disabled, as there's not much point in keeping the plugin around if its
disabled.

This means toggling the 'Play DRM Content' checkbox doesn't make the
Adobe CDM appear in the plugins list, and also if the Widevine CDM is
explicitly disabled, it still appears in the plugins list if 'Play DRM
Content' is checked, but it appears in a disabled state. As you'd expect.


MozReview-Commit-ID: Fq10vnhiaKO

UltraBlame original commit: 212d06d11b24ccf163c689233f382a5995aa9914