Bug 1295831 - VIDEO_INTER_KEYFRAME_MAX_MS==0 means only 1 keyframe - r=francois
Telemetry peer review through mozreview.

This patch is modifying this probe description sligthly, to add the collection
of a special value: '0' means that only one keyframe was found when playing a
video (played for at least enough time to be eligible for video-decode-suspend
feature.)

Please note that this will not modify existing data or their meaning: It will
just add this zero value (which is not otherwise possible or meaningful when
dealing with *inter-keyframe* timings), which will help gauge the relative
presence of these difficult-to-recover-from videos, compared to more "normal"
videos.

MozReview-Commit-ID: ANFtybW9BKu

UltraBlame original commit: 5576171f46151b38996fc96ff150ef311ddcc387