Bug 1259788 - Create haveEvents() and friends in head.js. r=jib
When we expect multiple events of the same type (in this case "addtrack" on
MediaStream) we can use this convenience function to attach a listener and
wait for the expected number of events before removing the listener and resolving.

This also adds a function to check that there were no extra events of the same
kind after the expected N events, and a function that checks that we don't see
a certain event within a given timeout period.

MozReview-Commit-ID: 1P8MkEI0Yzm

UltraBlame original commit: 38cf33e16b198fb2315217a8ccd91a42c4d2c5fb