Bug 1296929 - Match spec for col/colgroup.span, textarea.rows/cols; r=jst
The spec recently changed to match browsers better.  There's currently
not much interop in exact details of how this work.  This brings us in
line with the spec except for the limit of 1000 on the span attribute.

The added textarea failures are spurious, because I'm not updating our
local tests in this commit.  The new tests are submitted upstream at
<https://github.com/w3c/web-platform-tests/pull/3518>.

MozReview-Commit-ID: 1L8aUtF47Qi

UltraBlame original commit: 1442a6ab4296039c44f3e80a7a0b1aea6d74f970