Bug 1296532 - Fix MoofParser tests - r=jya
Well, this is embarrassing: The MoofParser test was not parsing anything,
because it was given an empty byte range!
Also HasMetadata() has to run before RebuildFragmentedIndex(), because the
latter moves the offset and would then leave nothing for the former to read.

MozReview-Commit-ID: ZB35lc8iaE

UltraBlame original commit: 4552565382c03eebdc01bb57b7c96e4f40864fad