Bug 1289432 - Drop the DispatchImmediate codepath. r=botond
Instead of having this special shortcut case that goes through MaybeHandleCurrentBlock
and DispatchImmediate, we can just add the input to the queue normally and call
ProcessInputBlocks() which effectively does the same thing. Doing it this way
also guarantees that mQueuedInputs is non-empty and mQueuedInputs[0] is the
input currently being processed. This is useful for future patches because it
will allow us to maintain the guarantee that the input block for the input
currently being processed can be readily accessed via GetCurrentInputBlock().

MozReview-Commit-ID: JuVbpsFLURo

UltraBlame original commit: b40d59513ee6954ab7fff73689fa834727b4a0fa