Bug 1302821 - Ensure owner has full privileges on directories; r=dustin
Previously, when recursively changing ownership on directories we would
only change the owner. We saw some permission denied failures in
automation where the new owner couldn't modify files or directories.
This *might* be due to the owner write bits not always being set. Or
it could be something else (such as a filesystem bug - *cough* AUFS
*cough*).

This commit changes our recursive chown implementation to ensure owner
read, write, and execute bits are set on directories.

Because we're now always calling stat(), the code for calling chown()
is made conditional because we have the stat information and can avoid
the extra system call if it would be a no-op.

MozReview-Commit-ID: JT9q3QR4Sit

UltraBlame original commit: 4f48a94c9d5a6cb169d0bcfe15aafebd65e3d30a