Bug 1302632: P5. Update mochitests. r=gerald
The tests expected that the error code would be MEDIA_ERR_DECODE whenever we fail to open a video. However, MEDIA_ERR_DECODE is to be used only when "An error of some description occurred while decoding the media resource, after the resource was established to be usable."

All those files have errors in their metadata. Which makes the resource unusable to start with.

Similarly, networkState would be set to  NETWORK_NO_SOURCE if the metadata couldn't be read.

MozReview-Commit-ID: KXVJmD3ZQlT

UltraBlame original commit: 3c4a5256eff7fea7dd49b13917b34ba0fcf80402